Implemented SetDefaults and SettingDefaults events.
Added Terraria.Main initialization to the SetupFixture; now some of the
tests are sort of integration tests, but that's okay given that, at some
point, we have to be coupled explicitly to Terraria types. Also removed
useCloud from SavingEventArgs because that parameter is impossible to use
properly. Finally, changed around the IDisposable implementation in
ServiceBase.